전체적인 네이밍 변경, 화면 최적화
원래의 ChatPage를 ChatRoom으로, 원래의 HomePage를 MainPage로, _buildUserList가 있던 것을 따로 분리해 ChatPage로 변경, 1번째 탭은 HomePage, 2번째 탭은 ChatPage, 3번째 탭은 SettingPage로 네이밍.
