+++ lib/pages/home_page.dart
@@ -1,15 +1,5 @@
-import 'dart:ui';
-
-import 'package:chat_app_tutorial/pages/chat_page.dart';
-import 'package:chat_app_tutorial/services/auth/auth_service.dart';
-import 'package:chat_app_tutorial/pages/Screen2.dart';
-import 'package:chat_app_tutorial/pages/Screen3.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:chat_app_tutorial/pages/loading_page.dart';
+
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -19,199 +9,20 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-  // instance of auth
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-  int _selectedIndex = 0;
-
-  // sign user out
-  void signOut() {
-    //get auth service
-    final authService = Provider.of<AuthService>(context, listen: false);
-
-    authService.signOut();
-  }
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        scrolledUnderElevation: 0,
-        title: IndexedStack(
-          index: _selectedIndex,
-          children: [
-            Row(
-              children: [
-                SizedBox(width: 15,),
-                Icon(Icons.cloud, color: Colors.lightBlue[200],size: 30,),
-                SizedBox(width: 5,),
-                Text("Home", style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold,)),
-              ],
-            ),
-            Row(
-              children: [
-                SizedBox(width: 15,),
-                Icon(Icons.cloud, color: Colors.lightBlue[200],size: 30,),
-                SizedBox(width: 5,),
-                Text("Chats", style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold,)),
-              ],
-            ),
-            Row(
-              children: [
-                SizedBox(width: 15,),
-                Icon(Icons.cloud, color: Colors.lightBlue[200],size: 30,),
-                SizedBox(width: 5,),
-                Text("Settings", style: TextStyle(color: Colors.black, fontWeight: FontWeight.bold,)),
-              ],
-            ),
-          ],
-        ),
-        actions: [
-          IndexedStack(
-            index: _selectedIndex,
-            children: [
-              // Page 1
-              Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Icon(Icons.add_box_outlined, color: Colors.black, size: 30, ),
-                ],
-              ),
-              // Page 2
-              Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  Icon(Icons.add_circle_outline_outlined, color: Colors.black, size: 30, ),
-                ],
-              ),
-              // Page 3
-              Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  IconButton(
-                    onPressed: signOut,
-                    icon: const Icon(Icons.logout,),
-                    color: Colors.black,
-                  ),
-                ],
-              ),
-            ],
-          )
-        ],
+    return const Scaffold(
 
-
-      ),
-      body: IndexedStack(
-        index: _selectedIndex,
-        children: [
-          _buildUserList(),
-          Screen2(),
-          Screen3()
-        ],
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        showSelectedLabels: false,
-        showUnselectedLabels: false,
-        selectedItemColor: Colors.lightBlueAccent[100],
-        iconSize: 27,
-        items: const <BottomNavigationBarItem>[
-          BottomNavigationBarItem(
-              icon: Icon(Icons.cloud_circle_outlined),
-              label: '탭1',
-              activeIcon: Icon(Icons.cloud_circle)
-          ),
-          BottomNavigationBarItem(
-              icon: Icon(Icons.chat_outlined),
-              label: '탭2',
-              activeIcon: Icon(Icons.chat)
-          ),
-          BottomNavigationBarItem(
-              icon: Icon(Icons.settings_outlined),
-              label: '탭3',
-              activeIcon: Icon(Icons.settings)
-          )
-        ],
-        currentIndex: _selectedIndex,
-        onTap: _onItemTapped,
-      ),
-    );
-  }
-/*
-  Widget Screen2() {
-    return Scaffold(
       body: Center(
         child: Text(
-          'Screen2 from local file',
+          'Screen2 from Screen2.dart',
           style: TextStyle(
             fontSize: 20.0,
           ),
+
         ),
       ),
     );
-  }
-
-  Widget Screen3() {
-    return Scaffold(
-      body: Center(
-        child: Text(
-          'Screen3 from local file',
-          style: TextStyle(
-            fontSize: 20.0,
-          ),
-        ),
-      ),
-    );
-  }
- */
-
-  // build a list of user except for the current Logged in user
-  Widget _buildUserList() {
-    return StreamBuilder<QuerySnapshot>(
-      stream: FirebaseFirestore.instance.collection('users').snapshots(),
-      builder: (context, snapshot) {
-        if (snapshot.hasError) {
-          return const Text("error");
-        }
-        if (snapshot.connectionState == ConnectionState.waiting) {
-          return LoadingPage();
-        }
-        return ListView(
-          children: snapshot.data!.docs
-              .map<Widget>((doc) => _buildUserListItem(doc))
-              .toList(),
-        );
-      },
-    );
-  }
-
-  // build individual user list items
-  Widget _buildUserListItem(DocumentSnapshot document) {
-    Map<String, dynamic> data = document.data()! as Map<String, dynamic>;
-
-    //display all users except for current user
-    if (_auth.currentUser!.email != data['name']) {
-      return ListTile(
-        title: Text(data['name']),
-        onTap: () {
-          //pass the clicked user's UID to the chat page
-          Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (context) => ChatPage(
-                        receiverUserEmail: data['name'],
-                        receiverUserID: data['uid'],
-                      )));
-        },
-      );
-    } else {
-      //return empty container
-      return Container();
-    }
   }
 }
