+++ lib/services/auth/auth_gate.dart
@@ -1,4 +1,4 @@
-import 'package:chat_app_tutorial/pages/home_page.dart';
+import 'package:chat_app_tutorial/pages/main_page.dart';
 import 'package:chat_app_tutorial/services/auth/login_or_register.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
@@ -14,7 +14,7 @@ class AuthGate extends StatelessWidget {
         builder: (context, snapshot){
           // user is logged in
           if(snapshot.hasData){
-            return const HomePage();
+            return const MainPage();
           }
           //user is not logged in
           else{
